SM-57: Create store endpoint for Reparation Request

+++ app/API/V1/Http/Controllers/ReparationRequestController.php
@@ -4,10 +4,14 @@ declare(strict_types=1);
 
 namespace App\API\V1\Http\Controllers;
 
+use App\API\V1\Http\Requests\StoreReparationRequest;
+use App\API\V1\Http\Requests\UpdateReparationRequest;
 use App\API\V1\Http\Resources\ReparationRequestCollection;
 use App\API\V1\Http\Resources\ReparationRequestResource;
+use App\Authentication\GuardsInterface;
 use App\Http\Controllers\Controller;
 use App\Models\ReparationRequest;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\JsonResponse;
 use Spatie\QueryBuilder\AllowedFilter;
 use Spatie\QueryBuilder\AllowedInclude;
@@ -16,7 +20,7 @@ use Spatie\QueryBuilder\QueryBuilder;
 final class ReparationRequestController extends Controller
 {
     /**
-     * @throws \Illuminate\Auth\Access\AuthorizationException
+     * @throws AuthorizationException
      */
     public function index(): ReparationRequestCollection
     {
@@ -43,13 +47,50 @@ final class ReparationRequestController extends Controller
         return new ReparationRequestCollection($reparationRequests);
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function show(ReparationRequest $reparationRequest): ReparationRequestResource
     {
+        $this->authorize('view', $reparationRequest);
+
         return new ReparationRequestResource($reparationRequest);
     }
 
+    /**
+     * @throws AuthorizationException
+     */
+    public function store(StoreReparationRequest $request): ReparationRequestResource
+    {
+        $this->authorize('create', ReparationRequest::class);
+
+        $reparationRequest = ReparationRequest::make($request->safe()->all());
+        $reparationRequest->reporter()->associate($request->user(GuardsInterface::REST_API));
+        $reparationRequest->save();
+
+        return new ReparationRequestResource($reparationRequest);
+    }
+
+    /**
+     * @throws AuthorizationException
+     */
+    public function update(UpdateReparationRequest $request, ReparationRequest $reparationRequest): ReparationRequestResource
+    {
+        $this->authorize('update', $reparationRequest);
+
+        $reparationRequest->fill($request->safe()->all());
+        $reparationRequest->save();
+
+        return new ReparationRequestResource($reparationRequest);
+    }
+
+    /**
+     * @throws AuthorizationException
+     */
     public function destroy(ReparationRequest $reparationRequest): JsonResponse
     {
+        $this->authorize('delete', $reparationRequest);
+
         $reparationRequest->delete();
 
         return new JsonResponse();
